agents/ai_analysis_agent/agent.py
"""
AI Analysis Agent for ADK Web UI
"""

from google.adk.agents import Agent

def analyze_campaign_data(campaign_data: str) -> dict:
    """Analyze campaign performance data and provide insights.
    
    Args:
        campaign_data: Campaign performance data to analyze
    
    Returns:
        dict: Analysis results with insights and recommendations
    """
    return {
        "analysis": f"Detailed analysis of {campaign_data}",
        "key_insights": [
            "Campaign performance is above industry average",
            "Click-through rate shows strong engagement",
            "Conversion rate has room for improvement",
            "Mobile traffic outperforms desktop"
        ],
        "recommendations": [
            "Increase budget allocation by 15%",
            "Test new ad copy variations",
            "Optimize landing pages for mobile",
            "Add negative keywords to reduce irrelevant clicks"
        ],
        "risk_assessment": "Low risk - campaign is performing well",
        "confidence_score": 0.85
    }

def generate_performance_insights(metrics: str, timeframe: str) -> dict:
    """Generate insights from performance metrics.
    
    Args:
        metrics: Performance metrics data
        timeframe: Time period for analysis
    
    Returns:
        dict: Performance insights and trends
    """
    return {
        "timeframe": timeframe,
        "metrics_analyzed": metrics,
        "trends": [
            "Steady growth in impressions over the period",
            "Cost per click decreased by 8%",
            "Quality score improved across keywords"
        ],
        "opportunities": [
            "Expand successful ad groups",
            "Implement automated bidding",
            "Test responsive search ads"
        ],
        "next_actions": [
            "Schedule A/B test for next week",
            "Review keyword performance",
            "Update ad extensions"
        ]
    }

def research_competitor_keywords(industry: str, competitor_names: list[str]) -> dict:
    """Research competitor keyword strategies.
    
    Args:
        industry: Industry or market segment
        competitor_names: List of competitor names to analyze
    
    Returns:
        dict: Competitor keyword analysis
    """
    return {
        "industry": industry,
        "competitors_analyzed": competitor_names,
        "keyword_gaps": [
            "long-tail commercial keywords",
            "brand + product combinations",
            "location-based keywords"
        ],
        "opportunities": [
            "Target competitor brand keywords",
            "Focus on underutilized long-tail terms",
            "Capitalize on seasonal keyword trends"
        ],
        "recommended_keywords": [
            f"{industry} solution",
            f"best {industry} software",
            f"{industry} comparison",
            f"affordable {industry} tools"
        ]
    }

root_agent = Agent(
    name="ai_analysis_agent",
    model="gemini-2.5-pro",
    description="AI analysis specialist for campaign data and performance insights",
    instruction="""
    You are an AI analysis specialist focused on Google Ads campaign optimization.
    
    Your expertise includes:
    - Campaign performance analysis and trend identification
    - Competitive intelligence and market research
    - Data-driven optimization recommendations
    - Risk assessment and opportunity identification
    
    Always provide:
    - Clear, actionable insights
    - Specific recommendations with expected impact
    - Risk assessments and confidence levels
    - Next steps for implementation
    
    Use the available tools to analyze data and generate comprehensive insights.
    """,
    tools=[analyze_campaign_data, generate_performance_insights, research_competitor_keywords]
)